+++ tests/Unit/Storage/Query/DBTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Projom\Tests\Unit\Storage;
+namespace Projom\Tests\Unit\Storage\Query;
 
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
